add web api

+++ BusinessLayer/Source/BLCertificateType.cs
@@ -3,6 +3,8 @@ using FRS.DatabaseContext;
 using FRS.Models;
 using System;
 using System.Collections.Generic;
+using System.Data.Common;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -18,7 +20,7 @@ namespace FRS.BusinessLayer
         public static List<CertificateType> GetAll()
         {
             List<CertificateType> list = null;
-            using (FamilyRelationshipContext dbContext = new FamilyRelationshipContext())
+            using (FamilyRelationshipContext dbContext = FamilyRelationshipContext.GetFamilyRelationshipContext())
             {
                 list = dbContext.CertificateType.ToList();               
             }
@@ -28,9 +30,9 @@ namespace FRS.BusinessLayer
         public static CertificateType Get(int CertificateTypeId)
         {
             CertificateType CertificateType = null;
-            using (FamilyRelationshipContext dbContext = new FamilyRelationshipContext())
+            using (FamilyRelationshipContext dbContext = FamilyRelationshipContext.GetFamilyRelationshipContext())
             {
-                CertificateType = dbContext.CertificateType.Where<CertificateType>(c=>c.CertificateTypeId == CertificateTypeId).FirstOrDefault();
+                CertificateType = dbContext.CertificateType.Find(CertificateTypeId);
             }
             return CertificateType;
         }
@@ -43,11 +45,11 @@ namespace FRS.BusinessLayer
         public static ErrorCode Add(CertificateType Member)
         {
             ErrorCode code = ErrorCode.Unknown_Error;
-            using (FamilyRelationshipContext dbContext = new FamilyRelationshipContext())
+            using (FamilyRelationshipContext dbContext = FamilyRelationshipContext.GetFamilyRelationshipContext())
             {
-                {
+                int count = dbContext.CertificateType.Where(c=>c.CertificateTypeName == Member.CertificateTypeName.Trim()).Count();
+                if (count > 0)
                     code = ErrorCode.DataAlreadyExist;
-                }
                 else
                 {
                     dbContext.CertificateType.Add(Member);
@@ -66,51 +68,54 @@ namespace FRS.BusinessLayer
         /// </summary>
         /// <param name="Member"></param>
         /// <returns></returns>
-        public static int Remove(CertificateType Member)
+        public static ErrorCode Remove(CertificateType Member)
         {
-            int rows = 0;
-            using (FamilyRelationshipContext dbContext = new FamilyRelationshipContext())
+            ErrorCode code = ErrorCode.Unknown_Error;
+            using (FamilyRelationshipContext dbContext = FamilyRelationshipContext.GetFamilyRelationshipContext())
             {
-                CertificateType rem = dbContext.CertificateType.Where<CertificateType>(c => c.CertificateTypeId == Member.CertificateTypeId).FirstOrDefault();
-                dbContext.CertificateType.Remove(rem);
-                rows = dbContext.SaveChanges();
+                CertificateType rem = dbContext.CertificateType.Find(Member.CertificateTypeId);
+                if (rem != null)
+                {
+                    dbContext.CertificateType.Remove(rem);
+                    int rows = dbContext.SaveChanges();
+                    if (rows > 0)
+                        code = ErrorCode.Success;
+                    else
+                        code = ErrorCode.DataDetectError;
+                }
+                else
+                    code = ErrorCode.Success;                
             }
-            return rows;
+            return code;
         }
 
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="CertificateTypeId"></param>
-        /// <returns></returns>
-        public static int Remove(int CertificateTypeId)
-        {
-            int rows = 0;
-            using (FamilyRelationshipContext dbContext = new FamilyRelationshipContext())
-            {
-                CertificateType rem = dbContext.CertificateType.Where<CertificateType>(c => c.CertificateTypeId == CertificateTypeId).FirstOrDefault();
-                dbContext.CertificateType.Remove(rem);
-                rows = dbContext.SaveChanges();
-            }
-            return rows;
-        }
 
         /// <summary>
         /// 
         /// </summary>
         /// <param name="Member"></param>
         /// <returns></returns>
-        public static int Update(CertificateType Member)
+        public static ErrorCode Modify(CertificateType Member)
         {
-            int rows = 0;
-            using (FamilyRelationshipContext dbContext = new FamilyRelationshipContext())
+            ErrorCode code = ErrorCode.Unknown_Error;
+            using (FamilyRelationshipContext dbContext = FamilyRelationshipContext.GetFamilyRelationshipContext())
             {
-                CertificateType up = dbContext.CertificateType.Where<CertificateType>(c => c.CertificateTypeId == Member.CertificateTypeId).FirstOrDefault();
-                up.CertificateTypeName = Member.CertificateTypeName;
-                up.Description = Member.Description;
-                rows = dbContext.SaveChanges();
+                CertificateType up = dbContext.CertificateType.Find(Member.CertificateTypeId);
+                if (up != null)
+                {
+                    up.CertificateTypeName = Member.CertificateTypeName;
+                    up.Description = Member.Description;
+                    int rows = dbContext.SaveChanges();
+
+                    if (rows > 0)
+                        code = ErrorCode.Success;
+                    else
+                        code = ErrorCode.DataModifyError;
+                }
+                else
+                    code = ErrorCode.DataNotExist;
             }
-            return rows;
+            return code;
         }
     }
 }
